moved script to pull everything together into own file
