+++ napkon_string_matching/__init__.py
@@ -1,3 +1,4 @@
 """
 Module containing scripts to match questions from different cohorts
 """
+from napkon_string_matching.matching import match
